% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mean.R
\name{my_mean}
\alias{my_mean}
\title{Sample mean}
\usage{
my_mean(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
sample mean
}
\description{
The function takes a vector and calculates the sample mean
}
\examples{
x <- 1:10
my_mean(x)
}
