% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_cup-data.R
\docType{data}
\name{worldcup_2010}
\alias{worldcup_2010}
\title{Summary statistics for world cup teams in 2010}
\format{A data frame with 32 rows and 15 variables:
\describe{
\item{Country}{The world cup team}
\item{\% Goals to shots}{Percentage of shots on goal that resulted in a goal}
...
}}
\source{
\url{jono}
}
\usage{
worldcup_2010
}
\description{
A dataset containing summary statistic for each team in
World Cup 2010.
}
\keyword{datasets}
